+++ subsys/ipc/ipc_service/backends/ipc_icmsg.h
@@ -3,11 +3,3 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-
-#include <zephyr/sys/spsc_pbuf.h>
-
-enum icmsg_state {
-	ICMSG_STATE_OFF,
-	ICMSG_STATE_BUSY,
-	ICMSG_STATE_READY
-};
